+++ src/tetra/js/tetra.core.js
@@ -121,12 +121,12 @@ const Tetra = {
         this.handleGroupNotify(payload);
         break;
 
-      case 'component.data_updated':
-        this.handleComponentUpdateData(payload);
+      case 'component.data_changed':
+        this.handleComponentDataChanged(payload);
         break;
 
       case 'component.removed':
-        this.handleComponentRemove(payload);
+        this.handleComponentRemoved(payload);
         break;
 
       case 'component.created':
@@ -192,7 +192,7 @@ const Tetra = {
       }
     }));
   },
-  handleComponentUpdateData(event) {
+  handleComponentDataChanged(event) {
     const { group, data, sender_id } = event;
     const components = this._get_components_by_subscribe_group(group);
     if (components.length === 0) {
@@ -223,7 +223,7 @@ const Tetra = {
       }
     });
   },
-  handleComponentRemove(event) {
+  handleComponentRemoved(event) {
     const { type, group, component_id, target_group, sender_id } = event;
 
     if (component_id) {